Refactor #743 - Move schedule list test resource generators to parent package
