+++ core/component/schedulelist-test/src/main/kotlin/com/nlab/reminder/core/component/schedulelist/ScheduleListResourceGenerator.kt
@@ -14,9 +14,8 @@
  * limitations under the License.
  */
 
-package com.nlab.reminder.core.component.schedulelist.content
+package com.nlab.reminder.core.component.schedulelist
 
-import com.nlab.reminder.core.component.schedulelist.ScheduleListResource
 import com.nlab.reminder.core.data.model.Link
 import com.nlab.reminder.core.data.model.LinkMetadata
 import com.nlab.reminder.core.data.model.Schedule
