+++ core/component/schedulelist-test/src/main/kotlin/com/nlab/reminder/core/component/schedulelist/UserScheduleListResourceGenerator.kt
@@ -14,10 +14,8 @@
  * limitations under the License.
  */
 
-package com.nlab.reminder.core.component.schedulelist.content
+package com.nlab.reminder.core.component.schedulelist
 
-import com.nlab.reminder.core.component.schedulelist.ScheduleListResource
-import com.nlab.reminder.core.component.schedulelist.UserScheduleListResource
 import com.nlab.reminder.core.data.model.ScheduleId
 import com.nlab.reminder.core.kotlin.collections.toSet
 import com.nlab.testkit.faker.genBoolean
