+++ feature/all/src/test/java/com/nlab/reminder/feature/all/GetUserScheduleListResourceReportFlowTest.kt
@@ -18,9 +18,9 @@ package com.nlab.reminder.feature.all
 
 import app.cash.turbine.test
 import com.nlab.reminder.core.component.schedulelist.GetUserScheduleListResourcesFlowUseCase
-import com.nlab.reminder.core.component.schedulelist.content.genScheduleListResource
-import com.nlab.reminder.core.component.schedulelist.content.genUserScheduleListResource
-import com.nlab.reminder.core.component.schedulelist.content.genUserScheduleListResources
+import com.nlab.reminder.core.component.schedulelist.genScheduleListResource
+import com.nlab.reminder.core.component.schedulelist.genUserScheduleListResource
+import com.nlab.reminder.core.component.schedulelist.genUserScheduleListResources
 import com.nlab.reminder.core.data.model.Schedule
 import com.nlab.reminder.core.data.model.ScheduleId
 import com.nlab.reminder.core.data.model.SchedulesLookup
